Use float for log value and chart value instead of integer

+++ src/db/model/chartData.ts
@@ -27,7 +27,7 @@ const chartDataSchema = {
     },
   },
   value: {
-    type: DataTypes.SMALLINT,
+    type: DataTypes.FLOAT,
     allowNull: false,
   },
 };
